fixed some bugs with administration
added sales as a feature to be looked at

+++ src/main/java/nl/iprwc/controller/GroupController.java
@@ -2,6 +2,7 @@ package nl.iprwc.controller;
 
 import nl.iprwc.db.GroupDAO;
 import nl.iprwc.exception.InvalidOperationException;
+import nl.iprwc.exception.NotFoundException;
 import nl.iprwc.model.Group;
 import nl.iprwc.model.User;
 
@@ -11,19 +12,22 @@ public class GroupController {
     private GroupDAO dao;
     private static GroupController instance;
 
-    public static synchronized GroupController getInstance() {
-        if (instance == null) {
-            instance = new GroupController();
-        }
-
+    public static GroupController getInstance() {
         return instance;
     }
+
+    static {
+        instance = new GroupController();
+    }
+
     private GroupController() {
         dao = new GroupDAO();
     }
 
-    public boolean deleteAccountGroup(String id) throws InvalidOperationException {
+    public boolean deleteAccountGroup(String id) throws InvalidOperationException, NotFoundException {
         try {
+            if (id == null) throw new NotFoundException();
+            System.out.println(id);
             return dao.deleteAccountGroup(id);
         } catch (SQLException | ClassNotFoundException e) {
             throw new InvalidOperationException();
